changed pathUtils.ts to utils.ts and added readDockerCommand and writeDockerCommand functions
